+++ src/frontend/components/PlayListModal/PlayListModal.jsx
@@ -3,15 +3,13 @@ import "./playlistmodal.css";
 import { GrClose, MdDeleteOutline } from "../../constants/react-icons";
 import { CreatePlaylist } from "../CreatePlaylist/CreatePlaylist";
 import { usePlaylists } from "../../context/playlist-context";
-import { deletePlaylist } from "../../services/playlist/deletePlaylist";
 import {
   ADD_TO_PLAYLIST,
   REMOVE_FROM_PLAYLIST,
   REMOVE_PLAYLIST,
 } from "../../constants/playlist-constants";
 import { useAuth } from "../../context/auth-context";
-import { addSpecificVideoToPlaylist } from "../../services/playlist/addSpecificVideoToPlaylist";
-import { deleteSpecificVideoFromPlaylist } from "../../services/playlist/deleteSpecificVideoFromPlaylist";
+import { deletePlaylist,removeVideoFromPlaylist,addVideoToPlaylist } from "../../services";
 
 export const PlayListModal = ({ closeFun, vid }) => {
   const { playlistDipatcher } = usePlaylists();
@@ -30,9 +28,9 @@ export const PlayListModal = ({ closeFun, vid }) => {
     playlistState: { playlists },
   } = usePlaylists();
 
-  const addVideoToPlaylist = async (id, vid) => {
+  const addVideoToPlaylistFun = async (id, vid) => {
     try {
-      const data = await addSpecificVideoToPlaylist(id, vid, tokenVal);
+      const data = await addVideoToPlaylist(id, vid, tokenVal);
       playlistDipatcher({ type: ADD_TO_PLAYLIST, payload: data.playlist });
     } catch (err) {
       console.log(err);
@@ -40,13 +38,13 @@ export const PlayListModal = ({ closeFun, vid }) => {
   };
   const deleteVideoFromPlaylist = async (id, vid_id) => {
     try {
-      const data = await deleteSpecificVideoFromPlaylist(id, vid_id, tokenVal);
+      const data = await removeVideoFromPlaylist(id, vid_id, tokenVal);
       playlistDipatcher({ type: REMOVE_FROM_PLAYLIST, payload: data.playlist });
     } catch (err) {
       console.log(err);
     }
   };
-  console.log(playlists);
+
   return (
     <div className="modal-container">
       <div className=" flex-col modal">
@@ -70,7 +68,7 @@ export const PlayListModal = ({ closeFun, vid }) => {
                     })}
                     onChange={(e) => {
                       if (e.target.checked) {
-                        addVideoToPlaylist(play._id, vid);
+                        addVideoToPlaylistFun(play._id, vid);
                       } else {
                         deleteVideoFromPlaylist(play._id, vid._id);
                       }
